requirements.txt
pandas>=2.0.0
numpy<2.0.0,>=1.26.0
scikit-surprise>=1.1.3
streamlit>=1.28.0
lightgbm>=4.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
requests>=2.31.0
